Fix KDevelop's detection for Visual Studio 2017 compiler and tools.
Visual Studio 2017 defines the VS150COMNTOOLS variable which kdevelop-msvc.bat ignored till now.

Also VS17 has a new folder layout, so vcvarsall.bat is under a different location. An if-else finds the script accordingly,
based on the registry. The script is less verbose in the cmd.

Last but not least, the script initializes the environment correctly, even if it isn't executed from within Developer's Command Prompt.